[MMCA-4298] Add Reads and Writes test for DateRange and some refactoring
